Use ez_setup instead of distribute_setup
Distribute merged into setuptools a while ago. Thus use the latter
again.
