+++ setup.py
@@ -6,7 +6,7 @@ import sys
 try:
     import setuptools
 except ImportError:
-    from distribute_setup import use_setuptools
+    from ez_setup import use_setuptools
     use_setuptools()
 from setuptools import setup, Extension
